fix:изменена функция "Пользователь"

+++ main-service/src/main/java/ru/practicum/mainservice/user/UserService.java
@@ -7,6 +7,7 @@ import ru.practicum.mainservice.user.dto.UserDto;
 import ru.practicum.mainservice.user.dto.UserSave;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -15,23 +16,28 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public List<UserDto> findAll(Integer from, Integer size, List<Integer> ids) {
+    public List<UserDto> findAll(final Integer from, final Integer size, final List<Integer> ids) {
         if (ids == null || ids.isEmpty()) {
             return userRepository.findAll(from, size);
         } else {
-            return userRepository.findAllById(ids);
+            return userRepository.findManyById(ids);
         }
     }
 
-    public UserDto create(UserSave userSave) {
-        UserDto userDtoSave = UserDto.builder()
+    public User create(final UserSave userSave) {
+        User user = User.builder()
                 .email(userSave.getEmail())
                 .name(userSave.getName())
                 .build();
-        return userRepository.save(userDtoSave);
+        return userRepository.save(user);
     }
 
-    public void delete(Long id) {
-        userRepository.deleteById(id);
+    public UserDto delete(final Long id) {
+        Optional<UserDto> user = userRepository.findOne(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return user.get();
+        }
+        return null;
     }
 }
